+++ CoCowork.DB/StoredProcedures/Payment/Payment_Update.sql
@@ -1,6 +1,6 @@
 ﻿CREATE PROC dbo.Payment_Update
 	@Id int,
-	@Amount int,
+	@Amount decimal(10,2),
 	@PaymentDate DateTime
 AS
 BEGIN
